+++ Changelog.md
@@ -8,8 +8,9 @@ API changes
 
 Enhancements
 
-* Support MKD	
-* Support RMD
+* Support MKD (make directory)
+* Support RMD (remove directory)
+* Support RNFR/RNTO (rename/move file/directory)
 
 ### 0.2.0